+++ test/mindconnect-agent-3-rsa.spec.ts
@@ -18,7 +18,7 @@ import {
 import { AgentManagementModels, MindSphereSdk } from "../src/api/sdk";
 import { decrypt, loadAuth, throwError } from "../src/api/utils";
 import { AgentUnitTestConfiguration, tearDownAgents, unitTestSetup } from "./test-agent-setup-utils";
-import {} from "./test-utils";
+import { getPasskeyForUnitTest } from "./test-utils";
 const log = debug("mindconnect-agent-test");
 chai.should();
 
@@ -26,7 +26,7 @@ describe("MindConnectApi Version 3 Agent (RSA_3072)", () => {
     const auth = loadAuth();
     const sdk = new MindSphereSdk({
         ...auth,
-        basicAuth: decrypt(auth, "passkey.4.unit.test"),
+        basicAuth: decrypt(auth, getPasskeyForUnitTest()),
     });
 
     let rsaConfig: IMindConnectConfiguration = ({} as unknown) as IMindConnectConfiguration;